% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{pdf_report}
\alias{pdf_report}
\title{WMF Product Analytics PDF Report}
\usage{
pdf_report(font_size = "12pt", page_margins = "1in",
  short_title = NULL, watermark = NULL, cite_r_packages = NULL,
  extra_bibs = NULL, toc = FALSE, lot = FALSE, lof = FALSE, ...)
}
\arguments{
\item{font_size}{base font size to typeset the document in (12pt default)}

\item{page_margins}{1in margins by default (recommended to keep unchanged)}

\item{short_title}{shorter version of the title to be printed on the
top-right corner of every page, except the front page}

\item{watermark}{watermark Text to be printed on the page, useful for first,
second, etc. drafts}

\item{cite_r_packages}{character vector of packages to generate a bibliography
for; necessary ones like \code{base} and \code{rmarkdown} will be cited regardless}

\item{extra_bibs}{character vector of paths to \code{.bib} files to include in
the final bibliography}

\item{toc}{T/F value for table of contents. See \code{?rmarkdown::pdf_document}
for details}

\item{lot, lof}{T/F value for list of tables/figures}

\item{...}{other options to be passed to \link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}
}
\description{
Creates a branded PDF document from R Markdown.
}
