% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmfpar.R
\docType{package}
\name{wmfpar-package}
\alias{wmfpar-package}
\alias{wmfpar}
\title{WMF Product Analytics reporting template}
\description{
A report template, based on \href{https://hebrewseniorlife.github.io/memor/}{memor}
template, for use by the
\href{https://www.mediawiki.org/wiki/Product_Analytics}{Product Analytics}
team within
\href{https://www.mediawiki.org/wiki/Wikimedia_Audiences}{Wikimedia Audiences}
department.
}
\details{
To be consistent with the documents put out by the Foundation, this
template requires the following fonts to be available in the system:
\itemize{
\item (Bold) \href{https://fonts.google.com/specimen/Montserrat}{Montserrat} for
header text.
\item \href{https://fonts.google.com/specimen/Source+Serif+Pro}{Source Serif Pro}
for body text.
}

Their use requires the availability of the XeLaTeX engine. The easiest way
to set this up is with \link[tinytex:install_tinytex]{tinytex::install_tinytex}. See
\href{https://yihui.name/tinytex/}{tinytex homepage} for more details on why
that's preferable to installing \href{http://www.tug.org/mactex/}{MacTeX}, for
example.

You also need \href{https://pandoc.org/}{pandoc}, which comes bundled with the
\href{https://www.rstudio.com/products/rstudio/}{RStudio IDE} so you don't need
to worry about that unless you're planning on rendering the
\href{https://rmarkdown.rstudio.com/}{R Markdown} document outside of RStudio.

\strong{Note}: data analysis does not have to be done in R in order to benefit
from this report template, as
\href{https://bookdown.org/yihui/rmarkdown/language-engines.html#python}{R Mardown supports Python chunks}
via \href{https://blog.rstudio.com/2018/03/26/reticulate-r-interface-to-python/}{reticulate}
package, which allows data to be transferred between R chunks and Python
chunks. See \href{https://rstudio.github.io/reticulate/articles/r_markdown.html}{this article}
for more information.
}
